requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
coverage==5.5
Django==3.2.5
django-jsoneditor==0.1.6
idna==3.2
isodate==0.6.0
jsonschema==3.2.0
openapi-schema-validator==0.1.5
packaging==21.0
postgres==3.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
